Let participatory space manifests register resources too
