+++ decidim-participatory_processes/lib/decidim/participatory_processes/participatory_space.rb
@@ -2,7 +2,6 @@
 
 Decidim.register_participatory_space(:participatory_processes) do |participatory_space|
   participatory_space.icon = "decidim/participatory_processes/icon.svg"
-  participatory_space.card = "decidim/participatory_processes/process"
   participatory_space.model_class_name = "Decidim::ParticipatoryProcess"
 
   participatory_space.participatory_spaces do |organization|
@@ -12,6 +11,11 @@ Decidim.register_participatory_space(:participatory_processes) do |participatory
   participatory_space.query_type = "Decidim::ParticipatoryProcesses::ParticipatoryProcessType"
 
   participatory_space.permissions_class_name = "Decidim::ParticipatoryProcesses::Permissions"
+
+  participatory_space.register_resource do |resource|
+    resource.model_class_name = "Decidim::ParticipatoryProcess"
+    resource.card = "decidim/participatory_processes/process"
+  end
 
   participatory_space.context(:public) do |context|
     context.engine = Decidim::ParticipatoryProcesses::Engine
